nat: extract static-NAT tests into tests_static.rs (#4409)
Move the 30 static (1:1) NAT tests -- static_nat_* covering host/block
translation, mapped-port, canonical CIDR-mask install, zone scoping and
the bidirectional reverse path -- verbatim into tests_static.rs, along
with their two test-local helpers block_snapshot() and
mapped_port_snapshot() (both used only by these tests, so no shared
helper is stranded).

Pure code motion: no test body edited, nothing renamed. nat/tests.rs
now holds 158 tests. See the first commit of this series for the split
rationale and module-wiring approach.
