+++ userspace-dp/src/nat/mod.rs
@@ -35,6 +35,9 @@ mod tests;
 #[cfg(test)]
 #[path = "tests_source.rs"]
 mod tests_source;
+#[cfg(test)]
+#[path = "tests_static.rs"]
+mod tests_static;
 
 /// #3096: per-flow interface / routing-instance identity passed into the NAT
 /// match path so an interface- or routing-instance-scoped rule-set matches
