+++ userspace-dp/src/nat/tests.rs
@@ -19,1093 +19,6 @@ use std::collections::BTreeSet;
 use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
 
 #[test]
-fn static_nat_dnat_matches_external_ip_v4() {
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            source_addresses: Vec::new(),
-            counter_id: 0,
-            name: "static-1".to_string(),
-            from_zone: "untrust".to_string(),
-            from_interface: String::new(),
-            from_routing_instance: String::new(),
-            external_ip: "203.0.113.10".to_string(),
-            internal_ip: "192.168.1.10".to_string(),
-            match_destination_port: 0,
-            mapped_port: 0,
-        }],
-        &crate::nat::NatCounterStore::default(),
-    );
-    let decision = table.match_dnat("203.0.113.10".parse().expect("ext"), "untrust");
-    assert_eq!(
-        decision,
-        Some(NatDecision {
-            rewrite_src: None,
-            rewrite_dst: Some("192.168.1.10".parse().expect("int")),
-            ..NatDecision::default()
-        })
-    );
-}
-
-#[test]
-fn static_nat_snat_matches_internal_ip_v4() {
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            source_addresses: Vec::new(),
-            counter_id: 0,
-            name: "static-1".to_string(),
-            from_zone: "trust".to_string(),
-            from_interface: String::new(),
-            from_routing_instance: String::new(),
-            external_ip: "203.0.113.10".to_string(),
-            internal_ip: "192.168.1.10".to_string(),
-            match_destination_port: 0,
-            mapped_port: 0,
-        }],
-        &crate::nat::NatCounterStore::default(),
-    );
-    let decision = table.match_snat("192.168.1.10".parse().expect("int"), "trust");
-    assert_eq!(
-        decision,
-        Some(NatDecision {
-            rewrite_src: Some("203.0.113.10".parse().expect("ext")),
-            rewrite_dst: None,
-            ..NatDecision::default()
-        })
-    );
-}
-
-#[test]
-fn static_nat_dnat_matches_external_ip_v6() {
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            source_addresses: Vec::new(),
-            counter_id: 0,
-            name: "static-v6".to_string(),
-            from_zone: "untrust".to_string(),
-            from_interface: String::new(),
-            from_routing_instance: String::new(),
-            external_ip: "2001:db8::1".to_string(),
-            internal_ip: "fd00::1".to_string(),
-            match_destination_port: 0,
-            mapped_port: 0,
-        }],
-        &crate::nat::NatCounterStore::default(),
-    );
-    let decision = table.match_dnat("2001:db8::1".parse().expect("ext"), "untrust");
-    assert_eq!(
-        decision,
-        Some(NatDecision {
-            rewrite_src: None,
-            rewrite_dst: Some("fd00::1".parse().expect("int")),
-            ..NatDecision::default()
-        })
-    );
-}
-
-#[test]
-fn static_nat_snat_matches_internal_ip_v6() {
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            source_addresses: Vec::new(),
-            counter_id: 0,
-            name: "static-v6".to_string(),
-            from_zone: "trust".to_string(),
-            from_interface: String::new(),
-            from_routing_instance: String::new(),
-            external_ip: "2001:db8::1".to_string(),
-            internal_ip: "fd00::1".to_string(),
-            match_destination_port: 0,
-            mapped_port: 0,
-        }],
-        &crate::nat::NatCounterStore::default(),
-    );
-    let decision = table.match_snat("fd00::1".parse().expect("int"), "trust");
-    assert_eq!(
-        decision,
-        Some(NatDecision {
-            rewrite_src: Some("2001:db8::1".parse().expect("ext")),
-            rewrite_dst: None,
-            ..NatDecision::default()
-        })
-    );
-}
-
-// #3031: block-to-block (subnet) static NAT — 1:1 by offset. A source
-// host maps to the same offset in the destination block (network bits
-// replaced, host bits preserved), both directions.
-
-fn block_snapshot(ext: &str, int: &str, from_zone: &str) -> StaticNATRuleSnapshot {
-    StaticNATRuleSnapshot {
-        source_addresses: Vec::new(),
-        counter_id: 0,
-        name: "block-1".to_string(),
-        from_zone: from_zone.to_string(),
-        from_interface: String::new(),
-        from_routing_instance: String::new(),
-        external_ip: ext.to_string(),
-        internal_ip: int.to_string(),
-        match_destination_port: 0,
-        mapped_port: 0,
-    }
-}
-
-#[test]
-fn static_nat_block_dnat_v4_preserves_offset() {
-    // 198.51.100.0/24 (external) -> 192.168.1.0/24 (internal). An inbound
-    // packet to 198.51.100.7 DNATs to 192.168.1.7 (offset .7 preserved).
-    let table = StaticNatTable::from_snapshots(
-        &[block_snapshot("198.51.100.0/24", "192.168.1.0/24", "untrust")],
-        &crate::nat::NatCounterStore::default(),
-    );
-    let decision = table.match_dnat("198.51.100.7".parse().expect("ext host"), "untrust");
-    assert_eq!(
-        decision,
-        Some(NatDecision {
-            rewrite_src: None,
-            rewrite_dst: Some("192.168.1.7".parse().expect("int host")),
-            ..NatDecision::default()
-        })
-    );
-}
-
-#[test]
-fn static_nat_block_snat_v4_reverses_offset() {
-    // Reverse: an outbound packet from 192.168.1.7 SNATs back to
-    // 198.51.100.7 (the inverse offset map), so return traffic matches.
-    let table = StaticNatTable::from_snapshots(
-        &[block_snapshot("198.51.100.0/24", "192.168.1.0/24", "untrust")],
-        &crate::nat::NatCounterStore::default(),
-    );
-    let decision = table.match_snat("192.168.1.7".parse().expect("int host"), "untrust");
-    assert_eq!(
-        decision,
-        Some(NatDecision {
-            rewrite_src: Some("198.51.100.7".parse().expect("ext host")),
-            rewrite_dst: None,
-            ..NatDecision::default()
-        })
-    );
-}
-
-#[test]
-fn static_nat_block_v6_preserves_offset_both_directions() {
-    // /120 -> /120 v6 block map. Offset ::7 preserved both directions.
-    let table = StaticNatTable::from_snapshots(
-        &[block_snapshot("2001:db8:a::/120", "fd00:1::/120", "untrust")],
-        &crate::nat::NatCounterStore::default(),
-    );
-    let dnat = table.match_dnat("2001:db8:a::7".parse().expect("ext v6"), "untrust");
-    assert_eq!(
-        dnat,
-        Some(NatDecision {
-            rewrite_src: None,
-            rewrite_dst: Some("fd00:1::7".parse().expect("int v6")),
-            ..NatDecision::default()
-        })
-    );
-    let snat = table.match_snat("fd00:1::7".parse().expect("int v6"), "untrust");
-    assert_eq!(
-        snat,
-        Some(NatDecision {
-            rewrite_src: Some("2001:db8:a::7".parse().expect("ext v6")),
-            rewrite_dst: None,
-            ..NatDecision::default()
-        })
-    );
-}
-
-#[test]
-fn static_nat_block_does_not_translate_outside_source_block() {
-    // An address OUTSIDE the external block is not DNAT'd, and an address
-    // outside the internal block is not SNAT'd.
-    let table = StaticNatTable::from_snapshots(
-        &[block_snapshot("198.51.100.0/24", "192.168.1.0/24", "untrust")],
-        &crate::nat::NatCounterStore::default(),
-    );
-    assert_eq!(
-        table.match_dnat("198.51.101.7".parse().expect("outside ext"), "untrust"),
-        None
-    );
-    assert_eq!(
-        table.match_snat("192.168.2.7".parse().expect("outside int"), "untrust"),
-        None
-    );
-}
-
-#[test]
-fn static_nat_block_mismatched_length_is_skipped() {
-    // A /24 -> /25 pair is not a 1:1 block map; the rule is skipped (no
-    // table entry, no translation), preserving the #2122 skip rationale.
-    let table = StaticNatTable::from_snapshots(
-        &[block_snapshot("198.51.100.0/24", "192.168.1.0/25", "untrust")],
-        &crate::nat::NatCounterStore::default(),
-    );
-    assert!(table.is_empty());
-    assert_eq!(
-        table.match_dnat("198.51.100.7".parse().expect("ext host"), "untrust"),
-        None
-    );
-}
-
-#[test]
-fn static_nat_block_with_port_is_dropped() {
-    // #3202: a block (subnet) pair that ALSO carries a port match / mapped-port
-    // is NOT representable by StaticNatBlock (address-only, all-port offset
-    // remap). Installing it would silently widen "port 80 of this /24 -> 8080"
-    // into "every port of the /24". The lenient-load backstop drops the rule
-    // (fail closed) instead of mis-installing it. The Go strict commit-check
-    // rejects this; this test pins the dataplane backstop.
-    //
-    // Fail-on-revert: removing the `snap.match_destination_port != 0 ||
-    // snap.mapped_port != 0` skip in from_snapshots installs an all-port block
-    // → table is non-empty and 198.51.100.7 DNATs to 192.168.1.7 → RED.
-    let mut snap = block_snapshot("198.51.100.0/24", "192.168.1.0/24", "untrust");
-    snap.match_destination_port = 80;
-    snap.mapped_port = 8080;
-    let table = StaticNatTable::from_snapshots(&[snap], &crate::nat::NatCounterStore::default());
-    assert!(
-        table.is_empty(),
-        "a block pair with a port mapping must be dropped, not installed as an all-port block"
-    );
-    assert_eq!(
-        table.match_dnat("198.51.100.7".parse().expect("ext host"), "untrust"),
-        None
-    );
-}
-
-#[test]
-fn static_nat_block_with_match_port_only_is_dropped() {
-    // A block pair with only a match destination-port (no mapped-port) is
-    // equally not representable — also dropped by the #3202 backstop.
-    let mut snap = block_snapshot("198.51.100.0/24", "192.168.1.0/24", "untrust");
-    snap.match_destination_port = 80;
-    let table = StaticNatTable::from_snapshots(&[snap], &crate::nat::NatCounterStore::default());
-    assert!(table.is_empty());
-}
-
-#[test]
-fn static_nat_host_v4_unchanged_with_block_support() {
-    // #3031 regression: a /32 host rule still behaves byte-identical to
-    // pre-#3031 (exact 1:1, no offset math). Both directions.
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            source_addresses: Vec::new(),
-            counter_id: 0,
-            name: "static-1".to_string(),
-            from_zone: "untrust".to_string(),
-            from_interface: String::new(),
-            from_routing_instance: String::new(),
-            external_ip: "203.0.113.10/32".to_string(),
-            internal_ip: "192.168.1.10/32".to_string(),
-            match_destination_port: 0,
-            mapped_port: 0,
-        }],
-        &crate::nat::NatCounterStore::default(),
-    );
-    assert_eq!(
-        table.match_dnat("203.0.113.10".parse().expect("ext"), "untrust"),
-        Some(NatDecision {
-            rewrite_src: None,
-            rewrite_dst: Some("192.168.1.10".parse().expect("int")),
-            ..NatDecision::default()
-        })
-    );
-    assert_eq!(
-        table.match_snat("192.168.1.10".parse().expect("int"), "untrust"),
-        Some(NatDecision {
-            rewrite_src: Some("203.0.113.10".parse().expect("ext")),
-            rewrite_dst: None,
-            ..NatDecision::default()
-        })
-    );
-    // A non-mapped host (not the /32) is not translated.
-    assert_eq!(
-        table.match_dnat("203.0.113.11".parse().expect("other"), "untrust"),
-        None
-    );
-}
-
-#[test]
-fn static_nat_zone_mismatch_returns_none_for_dnat() {
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            source_addresses: Vec::new(),
-            counter_id: 0,
-            name: "static-1".to_string(),
-            from_zone: "untrust".to_string(),
-            from_interface: String::new(),
-            from_routing_instance: String::new(),
-            external_ip: "203.0.113.10".to_string(),
-            internal_ip: "192.168.1.10".to_string(),
-            match_destination_port: 0,
-            mapped_port: 0,
-        }],
-        &crate::nat::NatCounterStore::default(),
-    );
-    // DNAT from wrong zone should fail
-    assert!(
-        table
-            .match_dnat("203.0.113.10".parse().expect("ext"), "trust")
-            .is_none()
-    );
-    // #2871: SNAT (reverse) now honors the EGRESS zone symmetrically with
-    // DNAT's ingress-zone gate. A packet egressing toward a zone OTHER than
-    // the rule's external `from zone` ("untrust") must NOT be source-NAT'd.
-    assert!(
-        table
-            .match_snat("192.168.1.10".parse().expect("int"), "dmz")
-            .is_none()
-    );
-}
-
-/// #2871 FAIL-ON-REVERT: static-NAT reverse (SNAT) must honor the EGRESS zone,
-/// mirroring the #2864 DNAT ingress-zone gate. An outbound packet sourced from
-/// a static-NAT internal IP but egressing toward a DIFFERENT internal zone
-/// (east-west) must NOT be source-translated to the public external IP — that
-/// was the cross-zone leak. It MUST still translate when egressing toward the
-/// rule's external `from zone`.
-///
-/// Drop the `egress_zone` gate from `match_snat_with_counter` (the pre-#2871
-/// "internal IP match is sufficient" behaviour) and the wrong-zone assertion
-/// below flips to Some(..) — this test goes RED.
-#[test]
-fn static_nat_snat_honors_egress_zone() {
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            source_addresses: Vec::new(),
-            counter_id: 0,
-            name: "web-server".to_string(),
-            // External zone of the rule: reverse SNAT applies only when the
-            // packet egresses toward THIS zone.
-            from_zone: "untrust".to_string(),
-            from_interface: String::new(),
-            from_routing_instance: String::new(),
-            external_ip: "203.0.113.10".to_string(),
-            internal_ip: "192.168.1.10".to_string(),
-            match_destination_port: 0,
-            mapped_port: 0,
-        }],
-        &crate::nat::NatCounterStore::default(),
-    );
-    let internal: IpAddr = "192.168.1.10".parse().expect("int");
-
-    // RIGHT zone: egressing toward the external zone "untrust" -> SNAT applies.
-    assert_eq!(
-        table.match_snat(internal, "untrust"),
-        Some(NatDecision {
-            rewrite_src: Some("203.0.113.10".parse().expect("ext")),
-            rewrite_dst: None,
-            ..NatDecision::default()
-        }),
-        "SNAT must apply when egressing toward the rule's external zone"
-    );
-
-    // WRONG zone: egressing toward another internal zone -> no source NAT.
-    // This is the assertion that fails without the #2871 egress-zone gate.
-    assert!(
-        table.match_snat(internal, "trust").is_none(),
-        "SNAT must NOT apply for an east-west packet egressing toward a \
-         non-external zone (cross-zone leak #2871)"
-    );
-    assert!(
-        table.match_snat(internal, "dmz").is_none(),
-        "SNAT must NOT apply when egressing toward an unrelated internal zone"
-    );
-}
-
-/// #2871: an empty external `from zone` ("any zone") still source-NATs on any
-/// egress zone — the egress-zone gate must not regress wildcard rules.
-#[test]
-fn static_nat_snat_empty_zone_matches_any_egress() {
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            source_addresses: Vec::new(),
-            counter_id: 0,
-            name: "wildcard".to_string(),
-            from_zone: String::new(),
-            from_interface: String::new(),
-            from_routing_instance: String::new(),
-            external_ip: "203.0.113.10".to_string(),
-            internal_ip: "192.168.1.10".to_string(),
-            match_destination_port: 0,
-            mapped_port: 0,
-        }],
-        &crate::nat::NatCounterStore::default(),
-    );
-    let internal: IpAddr = "192.168.1.10".parse().expect("int");
-    for egress in ["untrust", "trust", "dmz"] {
-        assert!(
-            table.match_snat(internal, egress).is_some(),
-            "wildcard (empty from_zone) SNAT must match any egress zone"
-        );
-    }
-}
-
-#[test]
-fn static_nat_empty_zone_matches_any() {
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            source_addresses: Vec::new(),
-            counter_id: 0,
-            name: "static-any".to_string(),
-            from_zone: String::new(),
-            from_interface: String::new(),
-            from_routing_instance: String::new(),
-            external_ip: "203.0.113.10".to_string(),
-            internal_ip: "192.168.1.10".to_string(),
-            match_destination_port: 0,
-            mapped_port: 0,
-        }],
-        &crate::nat::NatCounterStore::default(),
-    );
-    assert!(
-        table
-            .match_dnat("203.0.113.10".parse().expect("ext"), "untrust")
-            .is_some()
-    );
-    assert!(
-        table
-            .match_dnat("203.0.113.10".parse().expect("ext"), "trust")
-            .is_some()
-    );
-    assert!(
-        table
-            .match_snat("192.168.1.10".parse().expect("int"), "trust")
-            .is_some()
-    );
-}
-
-#[test]
-fn static_nat_bidirectional_reverse() {
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            source_addresses: Vec::new(),
-            counter_id: 0,
-            name: "static-1".to_string(),
-            from_zone: "untrust".to_string(),
-            from_interface: String::new(),
-            from_routing_instance: String::new(),
-            external_ip: "203.0.113.10".to_string(),
-            internal_ip: "192.168.1.10".to_string(),
-            match_destination_port: 0,
-            mapped_port: 0,
-        }],
-        &crate::nat::NatCounterStore::default(),
-    );
-    // Inbound DNAT: external -> internal
-    let dnat = table
-        .match_dnat("203.0.113.10".parse().expect("ext"), "untrust")
-        .expect("dnat");
-    assert_eq!(
-        dnat,
-        NatDecision {
-            rewrite_src: None,
-            rewrite_dst: Some("192.168.1.10".parse().expect("int")),
-            ..NatDecision::default()
-        }
-    );
-    // The reverse of DNAT should produce SNAT: on reply packets from
-    // the internal host, rewrite src back to the external IP.
-    // reverse().rewrite_src = self.rewrite_dst.map(|_| original_dst) = Some(external)
-    // reverse().rewrite_dst = self.rewrite_src.map(|_| original_src) = None
-    let original_src: IpAddr = "198.51.100.1".parse().expect("peer");
-    let original_dst: IpAddr = "203.0.113.10".parse().expect("ext");
-    let reverse = dnat.reverse(original_src, original_dst, 54321, 80);
-    assert_eq!(
-        reverse,
-        NatDecision {
-            rewrite_src: Some(original_dst),
-            rewrite_dst: None,
-            ..NatDecision::default()
-        }
-    );
-}
-
-#[test]
-fn static_nat_no_match_returns_none() {
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            source_addresses: Vec::new(),
-            counter_id: 0,
-            name: "static-1".to_string(),
-            from_zone: "untrust".to_string(),
-            from_interface: String::new(),
-            from_routing_instance: String::new(),
-            external_ip: "203.0.113.10".to_string(),
-            internal_ip: "192.168.1.10".to_string(),
-            match_destination_port: 0,
-            mapped_port: 0,
-        }],
-        &crate::nat::NatCounterStore::default(),
-    );
-    assert!(
-        table
-            .match_dnat("203.0.113.99".parse().expect("unknown"), "untrust")
-            .is_none()
-    );
-    assert!(
-        table
-            .match_snat("192.168.1.99".parse().expect("unknown"), "trust")
-            .is_none()
-    );
-}
-
-#[test]
-fn static_nat_invalid_ip_skipped() {
-    let table = StaticNatTable::from_snapshots(
-        &[
-            StaticNATRuleSnapshot {
-                source_addresses: Vec::new(),
-                counter_id: 0,
-                name: "bad".to_string(),
-                from_zone: String::new(),
-                from_interface: String::new(),
-                from_routing_instance: String::new(),
-                external_ip: "not-an-ip".to_string(),
-                internal_ip: "192.168.1.10".to_string(),
-                match_destination_port: 0,
-                mapped_port: 0,
-            },
-            StaticNATRuleSnapshot {
-                source_addresses: Vec::new(),
-                counter_id: 0,
-                name: "good".to_string(),
-                from_zone: String::new(),
-                from_interface: String::new(),
-                from_routing_instance: String::new(),
-                external_ip: "203.0.113.10".to_string(),
-                internal_ip: "192.168.1.10".to_string(),
-                match_destination_port: 0,
-                mapped_port: 0,
-            },
-        ],
-        &crate::nat::NatCounterStore::default(),
-    );
-    // The bad entry should be skipped, the good one should work
-    assert!(
-        table
-            .match_dnat("203.0.113.10".parse().expect("ext"), "any")
-            .is_some()
-    );
-}
-
-#[test]
-fn static_nat_external_ips_iterator() {
-    let table = StaticNatTable::from_snapshots(
-        &[
-            StaticNATRuleSnapshot {
-                source_addresses: Vec::new(),
-                counter_id: 0,
-                name: "s1".to_string(),
-                from_zone: String::new(),
-                from_interface: String::new(),
-                from_routing_instance: String::new(),
-                external_ip: "203.0.113.10".to_string(),
-                internal_ip: "192.168.1.10".to_string(),
-                match_destination_port: 0,
-                mapped_port: 0,
-            },
-            StaticNATRuleSnapshot {
-                source_addresses: Vec::new(),
-                counter_id: 0,
-                name: "s2".to_string(),
-                from_zone: String::new(),
-                from_interface: String::new(),
-                from_routing_instance: String::new(),
-                external_ip: "203.0.113.20".to_string(),
-                internal_ip: "192.168.1.20".to_string(),
-                match_destination_port: 0,
-                mapped_port: 0,
-            },
-        ],
-        &crate::nat::NatCounterStore::default(),
-    );
-    let mut ips: Vec<IpAddr> = table.external_ips().copied().collect();
-    ips.sort_by(|a, b| a.to_string().cmp(&b.to_string()));
-    assert_eq!(ips.len(), 2);
-    assert!(ips.contains(&"203.0.113.10".parse::<IpAddr>().unwrap()));
-    assert!(ips.contains(&"203.0.113.20".parse::<IpAddr>().unwrap()));
-}
-
-#[test]
-fn static_nat_canonical_cidr_mask_v4_installs_entry() {
-    // #2122: Junos emits static-NAT match/then in canonical prefix form
-    // ("203.0.113.5/32" / "10.0.0.5/32") and the Go compiler copies the mask
-    // verbatim into the snapshot. IpAddr::from_str rejects CIDR, so pre-fix
-    // every rule hit the Err/skip arm and was silently dropped — no
-    // translation occurred and the external IP was never registered as a
-    // local address. The parse must strip the /32 mask. This test FAILS on
-    // the unfixed code (table is empty, every assertion is None).
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            source_addresses: Vec::new(),
-            counter_id: 0,
-            name: "static-cidr".to_string(),
-            from_zone: "untrust".to_string(),
-            from_interface: String::new(),
-            from_routing_instance: String::new(),
-            external_ip: "203.0.113.5/32".to_string(),
-            internal_ip: "10.0.0.5/32".to_string(),
-            match_destination_port: 0,
-            mapped_port: 0,
-        }],
-        &crate::nat::NatCounterStore::default(),
-    );
-    // Inbound DNAT on the bare external IP -> bare internal IP.
-    assert_eq!(
-        table.match_dnat("203.0.113.5".parse().expect("ext"), "untrust"),
-        Some(NatDecision {
-            rewrite_src: None,
-            rewrite_dst: Some("10.0.0.5".parse().expect("int")),
-            ..NatDecision::default()
-        })
-    );
-    // Outbound SNAT on the bare internal IP -> bare external IP.
-    assert_eq!(
-        table.match_snat("10.0.0.5".parse().expect("int"), "untrust"),
-        Some(NatDecision {
-            rewrite_src: Some("203.0.113.5".parse().expect("ext")),
-            rewrite_dst: None,
-            ..NatDecision::default()
-        })
-    );
-    // The external IP must be registered (bare, no mask) for local delivery
-    // recognition (forwarding_build iterates external_ips()).
-    let ips: Vec<IpAddr> = table.external_ips().copied().collect();
-    assert_eq!(ips, vec!["203.0.113.5".parse::<IpAddr>().unwrap()]);
-}
-
-#[test]
-fn static_nat_canonical_cidr_mask_v6_installs_entry() {
-    // IPv6 canonical host form carries /128; same root cause as the v4 case.
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            source_addresses: Vec::new(),
-            counter_id: 0,
-            name: "static-cidr-v6".to_string(),
-            from_zone: "untrust".to_string(),
-            from_interface: String::new(),
-            from_routing_instance: String::new(),
-            external_ip: "2001:db8::1/128".to_string(),
-            internal_ip: "fd00::1/128".to_string(),
-            match_destination_port: 0,
-            mapped_port: 0,
-        }],
-        &crate::nat::NatCounterStore::default(),
-    );
-    assert_eq!(
-        table.match_dnat("2001:db8::1".parse().expect("ext"), "untrust"),
-        Some(NatDecision {
-            rewrite_src: None,
-            rewrite_dst: Some("fd00::1".parse().expect("int")),
-            ..NatDecision::default()
-        })
-    );
-    assert_eq!(
-        table.match_snat("fd00::1".parse().expect("int"), "untrust"),
-        Some(NatDecision {
-            rewrite_src: Some("2001:db8::1".parse().expect("ext")),
-            rewrite_dst: None,
-            ..NatDecision::default()
-        })
-    );
-    let ips: Vec<IpAddr> = table.external_ips().copied().collect();
-    assert_eq!(ips, vec!["2001:db8::1".parse::<IpAddr>().unwrap()]);
-}
-
-#[test]
-fn static_nat_cidr_and_bare_coexist() {
-    // A masked rule and a bare-IP rule must both install — stripping the mask
-    // on one must not regress the other.
-    let table = StaticNatTable::from_snapshots(
-        &[
-            StaticNATRuleSnapshot {
-                source_addresses: Vec::new(),
-                counter_id: 0,
-                name: "masked".to_string(),
-                from_zone: String::new(),
-                from_interface: String::new(),
-                from_routing_instance: String::new(),
-                external_ip: "203.0.113.5/32".to_string(),
-                internal_ip: "10.0.0.5".to_string(),
-                match_destination_port: 0,
-                mapped_port: 0,
-            },
-            StaticNATRuleSnapshot {
-                source_addresses: Vec::new(),
-                counter_id: 0,
-                name: "bare".to_string(),
-                from_zone: String::new(),
-                from_interface: String::new(),
-                from_routing_instance: String::new(),
-                external_ip: "203.0.113.6".to_string(),
-                internal_ip: "10.0.0.6/32".to_string(),
-                match_destination_port: 0,
-                mapped_port: 0,
-            },
-        ],
-        &crate::nat::NatCounterStore::default(),
-    );
-    assert!(
-        table
-            .match_dnat("203.0.113.5".parse().expect("ext"), "any")
-            .is_some()
-    );
-    assert!(
-        table
-            .match_dnat("203.0.113.6".parse().expect("ext"), "any")
-            .is_some()
-    );
-    assert!(
-        table
-            .match_snat("10.0.0.5".parse().expect("int"), "any")
-            .is_some()
-    );
-    assert!(
-        table
-            .match_snat("10.0.0.6".parse().expect("int"), "any")
-            .is_some()
-    );
-}
-
-#[test]
-fn static_nat_non_host_mask_rejected() {
-    // A non-host mask (or garbage suffix) is NOT a canonical host form and
-    // must be rejected (skipped), not silently coerced to a host route — it
-    // would otherwise translate the wrong scope. Pre-#2122 ALL of these were
-    // rejected too (the bare parser errored on any mask), so this is not a
-    // regression; it keeps the hardened parse strict about misconfiguration.
-    for bad in [
-        "203.0.113.5/24",      // non-host v4 prefix
-        "203.0.113.5/notanum", // non-numeric mask
-        "203.0.113.5/",        // empty mask
-        "203.0.113.5//32",     // double slash
-        "203.0.113.5/128",     // v6 host length applied to a v4 address
-        "2001:db8::1/64",      // non-host v6 prefix
-        "2001:db8::1/32",      // v4 host length applied to a v6 address
-    ] {
-        let table = StaticNatTable::from_snapshots(
-            &[StaticNATRuleSnapshot {
-                source_addresses: Vec::new(),
-                counter_id: 0,
-                name: "bad-mask".to_string(),
-                from_zone: String::new(),
-                from_interface: String::new(),
-                from_routing_instance: String::new(),
-                external_ip: bad.to_string(),
-                internal_ip: "10.0.0.5".to_string(),
-                match_destination_port: 0,
-                mapped_port: 0,
-            }],
-            &crate::nat::NatCounterStore::default(),
-        );
-        assert_eq!(
-            table.external_ips().count(),
-            0,
-            "non-host/garbage mask {bad:?} must be rejected, not installed"
-        );
-    }
-}
-
-// --- #2491 static-NAT port / mapped-port forwarding ---
-
-fn mapped_port_snapshot() -> StaticNATRuleSnapshot {
-    StaticNATRuleSnapshot {
-        source_addresses: Vec::new(),
-        counter_id: 0,
-        name: "port-map".to_string(),
-        from_zone: "untrust".to_string(),
-        from_interface: String::new(),
-        from_routing_instance: String::new(),
-        external_ip: "203.0.113.1/32".to_string(),
-        internal_ip: "10.0.0.5/32".to_string(),
-        match_destination_port: 8080,
-        mapped_port: 80,
-    }
-}
-
-// Forward (inbound DNAT): a packet to the external IP on the matched external
-// port is translated to the internal IP AND the destination port is rewritten
-// to the mapped-port. A packet to the same IP on a DIFFERENT port misses (no
-// port-less fallback entry, so no translation).
-//
-// Fail-on-revert: dropping `rewrite_dst_port: entry.mapped_port` in
-// match_dnat_with_counter turns the rewrite_dst_port assertion RED.
-#[test]
-fn static_nat_mapped_port_dnat_rewrites_dst_port() {
-    let table =
-        StaticNatTable::from_snapshots(&[mapped_port_snapshot()], &crate::nat::NatCounterStore::default());
-    let ext: IpAddr = "203.0.113.1".parse().unwrap();
-    let int: IpAddr = "10.0.0.5".parse().unwrap();
-
-    let (decision, _) = table
-        .match_dnat_with_counter(ext, 8080, "untrust")
-        .expect("matched external port");
-    assert_eq!(decision.rewrite_dst, Some(int), "dst IP must be the internal host");
-    assert_eq!(decision.rewrite_dst_port, Some(80), "dst port must be the mapped-port");
-
-    // A non-matching external port must NOT translate (no whole-address
-    // fallback present).
-    assert!(
-        table.match_dnat_with_counter(ext, 9999, "untrust").is_none(),
-        "non-matching external port must miss"
-    );
-}
-
-// Reverse (outbound return SNAT): the return packet from the internal host's
-// mapped-port has its source IP rewritten back to the external IP AND its
-// source port un-translated back to the external (match) port.
-//
-// Fail-on-revert: dropping `rewrite_src_port` in match_snat_with_counter (or
-// keying snat on the internal IP without the mapped-port) turns the
-// rewrite_src_port assertion RED / the lookup miss.
-#[test]
-fn static_nat_mapped_port_snat_untranslates_src_port() {
-    let table =
-        StaticNatTable::from_snapshots(&[mapped_port_snapshot()], &crate::nat::NatCounterStore::default());
-    let ext: IpAddr = "203.0.113.1".parse().unwrap();
-    let int: IpAddr = "10.0.0.5".parse().unwrap();
-
-    let (decision, _) = table
-        .match_snat_with_counter(int, 80, "untrust")
-        .expect("matched internal mapped-port");
-    assert_eq!(decision.rewrite_src, Some(ext), "src IP must be the external IP");
-    assert_eq!(
-        decision.rewrite_src_port,
-        Some(8080),
-        "src port must be un-translated to the external match port"
-    );
-}
-
-// A whole-address rule and a port-mapped rule can coexist on the SAME external
-// IP: the port-mapped entry wins on its exact port; everything else falls back
-// to the whole-address mapping (no port rewrite).
-#[test]
-fn static_nat_port_mapped_and_whole_address_coexist() {
-    let table = StaticNatTable::from_snapshots(
-        &[
-            mapped_port_snapshot(), // 203.0.113.1:8080 -> 10.0.0.5:80
-            StaticNATRuleSnapshot {
-                source_addresses: Vec::new(),
-                counter_id: 0,
-                name: "whole".to_string(),
-                from_zone: "untrust".to_string(),
-                from_interface: String::new(),
-                from_routing_instance: String::new(),
-                external_ip: "203.0.113.1/32".to_string(),
-                internal_ip: "10.0.0.9/32".to_string(),
-                match_destination_port: 0,
-                mapped_port: 0,
-            },
-        ],
-        &crate::nat::NatCounterStore::default(),
-    );
-    let ext: IpAddr = "203.0.113.1".parse().unwrap();
-
-    // Port-mapped entry wins on 8080.
-    let (port_dec, _) = table.match_dnat_with_counter(ext, 8080, "untrust").expect("port match");
-    assert_eq!(port_dec.rewrite_dst, Some("10.0.0.5".parse().unwrap()));
-    assert_eq!(port_dec.rewrite_dst_port, Some(80));
-
-    // Any other port falls back to the whole-address mapping (no port rewrite).
-    let (whole_dec, _) = table.match_dnat_with_counter(ext, 443, "untrust").expect("fallback match");
-    assert_eq!(whole_dec.rewrite_dst, Some("10.0.0.9".parse().unwrap()));
-    assert_eq!(whole_dec.rewrite_dst_port, None);
-}
-
-// Fail-closed: a mapped_port without a match_destination_port (which the Go
-// strict commit-check rejects, but can slip through the lenient load path) is
-// demoted to a whole-address 1:1 — no port rewrite, no orphaned port key.
-#[test]
-fn static_nat_mapped_port_without_match_port_demotes_to_whole_address() {
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            source_addresses: Vec::new(),
-            counter_id: 0,
-            name: "orphan".to_string(),
-            from_zone: "untrust".to_string(),
-            from_interface: String::new(),
-            from_routing_instance: String::new(),
-            external_ip: "203.0.113.1/32".to_string(),
-            internal_ip: "10.0.0.5/32".to_string(),
-            match_destination_port: 0,
-            mapped_port: 80,
-        }],
-        &crate::nat::NatCounterStore::default(),
-    );
-    let ext: IpAddr = "203.0.113.1".parse().unwrap();
-    // Matches as a whole-address entry on any port, with NO port rewrite.
-    let (decision, _) = table.match_dnat_with_counter(ext, 12345, "untrust").expect("whole-address match");
-    assert_eq!(decision.rewrite_dst, Some("10.0.0.5".parse().unwrap()));
-    assert_eq!(decision.rewrite_dst_port, None, "orphaned mapped-port must not rewrite");
-}
-
-// #2769: a `match destination-port` WITHOUT a `mapped-port` is a port-scoped
-// 1:1 (no port translation). The reverse SNAT MUST stay scoped to the matched
-// port — the internal service runs on, and its return packets leave from, the
-// matched port (no translation happened). A return packet from ANY OTHER
-// source port on the internal host MUST NOT be source-translated.
-//
-// Before #2769 the SNAT entry was keyed on `(internal_ip, None)`, so the
-// reverse SNAT matched every source port → it source-translated every service
-// on the internal host, not just the one port-scoped inbound. This is the
-// whole-host NAT broadening the issue reports.
-//
-// Fail-on-revert: reverting the `let snat_port = mapped_port.or(match_dst_port)`
-// scoping in from_snapshots back to `let snat_port = mapped_port` re-keys the
-// SNAT entry on `None`; the off-port lookup then matches the fallback entry and
-// the `is_none()` assertion goes RED.
-#[test]
-fn static_nat_match_port_without_mapped_port_scopes_reverse_snat() {
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            source_addresses: Vec::new(),
-            counter_id: 0,
-            name: "port-scoped-1to1".to_string(),
-            from_zone: "untrust".to_string(),
-            from_interface: String::new(),
-            from_routing_instance: String::new(),
-            external_ip: "203.0.113.1/32".to_string(),
-            internal_ip: "10.0.0.5/32".to_string(),
-            match_destination_port: 8080,
-            mapped_port: 0,
-        }],
-        &crate::nat::NatCounterStore::default(),
-    );
-    let ext: IpAddr = "203.0.113.1".parse().unwrap();
-    let int: IpAddr = "10.0.0.5".parse().unwrap();
-
-    // Inbound DNAT is scoped to the matched port (no port-less fallback).
-    let (dnat, _) = table
-        .match_dnat_with_counter(ext, 8080, "untrust")
-        .expect("matched external port");
-    assert_eq!(dnat.rewrite_dst, Some(int), "dst IP must be the internal host");
-    assert_eq!(
-        dnat.rewrite_dst_port, None,
-        "no mapped-port: destination port is not rewritten"
-    );
-    assert!(
-        table.match_dnat_with_counter(ext, 9999, "untrust").is_none(),
-        "off-port inbound must miss (port-scoped, no whole-address fallback)"
-    );
-
-    // Reverse SNAT: a return packet leaving the internal host on the matched
-    // port (8080 — no translation happened) IS source-translated to the
-    // external IP, with NO source-port rewrite.
-    let (snat, _) = table
-        .match_snat_with_counter(int, 8080, "untrust")
-        .expect("return from the port-scoped service");
-    assert_eq!(snat.rewrite_src, Some(ext), "src IP must be the external IP");
-    assert_eq!(
-        snat.rewrite_src_port, None,
-        "no port translation: source port is not rewritten"
-    );
-
-    // The bug: a packet from ANY OTHER source port on the internal host MUST
-    // NOT be source-translated. This is the whole-host broadening #2769 fixes.
-    assert!(
-        table.match_snat_with_counter(int, 1234, "untrust").is_none(),
-        "off-port outbound must NOT be source-translated (reverse SNAT must \
-         stay scoped to the matched port)"
-    );
-}
-
-// #2864: a port-specific static-NAT DNAT entry whose `from_zone` does NOT match
-// the packet's ingress zone MUST NOT short-circuit the lookup to `None`. It must
-// fall through to the whole-address `(dst_ip, None)` entry, which carries its own
-// (matching/empty) zone constraint. Before #2864 the zone check ran ONCE after
-// the `or_else()` precedence resolved, so a zone-mismatched port-specific entry
-// returned `None` and the whole-address rule was silently bypassed.
-//
-// Fail-on-revert: removing the per-candidate `.filter(zone_ok)` fall-through in
-// match_dnat_with_counter (so a port-specific zone-fail short-circuits to None)
-// turns the `fallback match` expect RED.
-#[test]
-fn static_nat_dnat_port_zone_mismatch_falls_back_to_whole_address() {
-    let ext: IpAddr = "203.0.113.1".parse().unwrap();
-    let table = StaticNatTable::from_snapshots(
-        &[
-            // Port-specific entry constrained to a DIFFERENT zone than the
-            // packet will arrive on. (203.0.113.1:8080 from "dmz" -> 10.0.0.5:80)
-            StaticNATRuleSnapshot {
-                source_addresses: Vec::new(),
-                counter_id: 0,
-                name: "port-dmz".to_string(),
-                from_zone: "dmz".to_string(),
-                from_interface: String::new(),
-                from_routing_instance: String::new(),
-                external_ip: "203.0.113.1/32".to_string(),
-                internal_ip: "10.0.0.5/32".to_string(),
-                match_destination_port: 8080,
-                mapped_port: 80,
-            },
-            // Whole-address entry valid for the packet's actual ingress zone.
-            StaticNATRuleSnapshot {
-                source_addresses: Vec::new(),
-                counter_id: 0,
-                name: "whole-untrust".to_string(),
-                from_zone: "untrust".to_string(),
-                from_interface: String::new(),
-                from_routing_instance: String::new(),
-                external_ip: "203.0.113.1/32".to_string(),
-                internal_ip: "10.0.0.9/32".to_string(),
-                match_destination_port: 0,
-                mapped_port: 0,
-            },
-        ],
-        &crate::nat::NatCounterStore::default(),
-    );
-
-    // A packet to the port-specific port (8080) but from the WRONG zone for the
-    // port entry ("untrust", not "dmz") must NOT short-circuit: it falls back to
-    // the whole-address entry, whose zone ("untrust") matches.
-    let (whole_dec, _) = table
-        .match_dnat_with_counter(ext, 8080, "untrust")
-        .expect("fallback match — port-specific zone-fail must fall through to whole-address");
-    assert_eq!(
-        whole_dec.rewrite_dst,
-        Some("10.0.0.9".parse().unwrap()),
-        "must translate via the whole-address entry, not the zone-mismatched port entry"
-    );
-    assert_eq!(
-        whole_dec.rewrite_dst_port, None,
-        "whole-address entry has no mapped-port"
-    );
-
-    // Port-specific precedence is preserved when its zone DOES match: a packet
-    // to 8080 from "dmz" hits the port entry (mapped-port rewrite), NOT the
-    // whole-address entry.
-    let (port_dec, _) = table
-        .match_dnat_with_counter(ext, 8080, "dmz")
-        .expect("port-specific entry wins when its zone matches");
-    assert_eq!(port_dec.rewrite_dst, Some("10.0.0.5".parse().unwrap()));
-    assert_eq!(port_dec.rewrite_dst_port, Some(80));
-
-    // A non-port packet from "dmz" — the port entry does not key, and the
-    // whole-address entry's zone ("untrust") does not match "dmz" → no DNAT.
-    assert!(
-        table.match_dnat_with_counter(ext, 443, "dmz").is_none(),
-        "no candidate matches the ingress zone → no DNAT"
-    );
-
-    // A non-port packet from "untrust" hits the whole-address entry.
-    let (any_dec, _) = table
-        .match_dnat_with_counter(ext, 443, "untrust")
-        .expect("whole-address entry matches its own zone on any port");
-    assert_eq!(any_dec.rewrite_dst, Some("10.0.0.9".parse().unwrap()));
-
-    // No entry for an unknown IP → no DNAT.
-    assert!(
-        table
-            .match_dnat_with_counter("203.0.113.250".parse().unwrap(), 8080, "untrust")
-            .is_none(),
-        "unknown destination IP → no DNAT"
-    );
-}
-
-// --- DNAT table tests ---
-
-#[test]
 fn dnat_basic_lookup_tcp() {
     let table = DnatTable::from_snapshots(
         &[DestinationNATRuleSnapshot {
